ENH: Move filter tools from utils to rtprocessing library
