+++ applications/mne_analyze/plugins/rawdataviewer/fiffrawview.h
@@ -67,7 +67,7 @@ namespace ANSHAREDLIB {
     class FiffRawViewModel;
 }
 
-namespace UTILSLIB {
+namespace RTPROCESSINGLIB {
     class FilterData;
 }
 
@@ -237,7 +237,7 @@ public:
      *
      * @param[in] filterData   the currently active filter
      */
-    void setFilter(const UTILSLIB::FilterData &filterData);
+    void setFilter(const RTPROCESSINGLIB::FilterData &filterData);
 
     //=========================================================================================================
     /**
