+++ applications/mne_scan/libs/scDisp/realtimemultisamplearraywidget.cpp
@@ -54,7 +54,7 @@
 
 #include <scMeas/realtimemultisamplearray.h>
 
-#include <utils/filterTools/filterdata.h>
+#include <rtprocessing/helpers/filterdata.h>
 
 //=============================================================================================================
 // QT INCLUDES
@@ -74,7 +74,7 @@
 using namespace SCDISPLIB;
 using namespace SCMEASLIB;
 using namespace DISPLIB;
-using namespace UTILSLIB;
+using namespace RTPROCESSINGLIB;
 
 //=============================================================================================================
 // DEFINE MEMBER METHODS
