+++ libraries/disp/viewers/helpers/filterplotscene.cpp
@@ -56,7 +56,7 @@
 
 using namespace DISPLIB;
 using namespace Eigen;
-using namespace UTILSLIB;
+using namespace RTPROCESSINGLIB;
 
 //=============================================================================================================
 // DEFINE MEMBER METHODS
