+++ libraries/disp/viewers/helpers/rtfiffrawviewmodel.h
@@ -44,7 +44,8 @@
 
 #include <fiff/fiff_types.h>
 #include <fiff/fiff_proj.h>
-#include <utils/filterTools/filterdata.h>
+
+#include <rtprocessing/helpers/filterdata.h>
 
 //=============================================================================================================
 // QT INCLUDES
@@ -67,10 +68,6 @@
 
 namespace FIFFLIB {
     class FiffInfo;
-}
-
-namespace UTILSLIB {
-    class FilterData;
 }
 
 //=============================================================================================================
@@ -332,7 +329,7 @@ public:
      *
      * @param[in] filterData    list of the new filter
      */
-    void setFilter(QList<UTILSLIB::FilterData> filterData);
+    void setFilter(QList<RTPROCESSINGLIB::FilterData> filterData);
 
     //=========================================================================================================
     /**
@@ -512,7 +509,7 @@ private:
      */
     void initSphara();
 
-    static void doFilterPerChannelRTMSA(QPair<QList<UTILSLIB::FilterData>,QPair<int,Eigen::RowVectorXd> > &channelDataTime);
+    static void doFilterPerChannelRTMSA(QPair<QList<RTPROCESSINGLIB::FilterData>,QPair<int,Eigen::RowVectorXd> > &channelDataTime);
 
     //=========================================================================================================
     /**
@@ -598,7 +595,7 @@ private:
     QMap<int,QList<QPair<int,double> > >m_qMapDetectedTriggerOld;                   /**< Old detected trigger for each trigger channel. */
     QMap<int,QList<QPair<int,double> > >m_qMapDetectedTriggerOldFreeze;             /**< Old detected trigger for each trigger channel while display is freezed. */
     QMap<qint32,float>                  m_qMapChScaling;                            /**< Channel scaling map. */
-    QList<UTILSLIB::FilterData>         m_filterKernel;                               /**< List of currently active filters. */
+    QList<RTPROCESSINGLIB::FilterData>  m_filterKernel;                             /**< List of currently active filters. */
     QStringList                         m_filterChannelList;                        /**< List of channels which are to be filtered.*/
     QStringList                         m_visibleChannelList;                       /**< List of currently visible channels in the view.*/
     QMap<qint32,qint32>                 m_qMapIdxRowSelection;                      /**< Selection mapping.*/
