+++ libraries/rtprocessing/helpers/parksmcclellan.cpp
@@ -75,7 +75,7 @@
 // USED NAMESPACES
 //=============================================================================================================
 
-using namespace UTILSLIB;
+using namespace RTPROCESSINGLIB;
 using namespace Eigen;
 
 //=============================================================================================================
